new file:   src/assign/trail_saving.rs
