+++ src/assign/propagate.rs
@@ -2,10 +2,7 @@
 /// This version can handle Chronological and Non Chronological Backtrack.
 use {
     super::{AssignIF, AssignStack, VarHeapIF, VarManipulateIF},
-    crate::{
-        cdb::{self, ClauseDBIF},
-        types::*,
-    },
+    crate::{cdb::ClauseDBIF, types::*},
 };
 
 /// API for Boolean Constraint Propagation like [`propagate`](`crate::assign::PropagateIF::propagate`), [`assign_by_decision`](`crate::assign::PropagateIF::assign_by_decision`), [`cancel_until`](`crate::assign::PropagateIF::cancel_until`), and so on.
@@ -201,26 +198,15 @@ impl PropagateIF for AssignStack {
             self.save_best_phases();
             self.best_assign = false;
         }
+
+        #[cfg(feature = "chrono_BT")]
+        let mut unpropagated: Vec<Lit> = Vec::new();
+
         // We assume that backtrack is never happened in level zero.
         let lim = self.trail_lim[lv as usize];
 
         #[cfg(feature = "trail_saving")]
-        {
-            let dl = self.trail_lim.len();
-            self.clear_saved_trail();
-            if 2 <= dl {
-                let lim2 = self.trail_lim[dl - 2];
-                for i in (lim..lim2).rev() {
-                    let l = self.trail[i];
-                    let vi = l.vi();
-                    self.trail_saved.push(l);
-                    self.reason_saved[vi] = self.reason[vi];
-                }
-            }
-        }
-
-        #[cfg(feature = "chrono_BT")]
-        let mut unpropagated: Vec<Lit> = Vec::new();
+        self.save_trail(lim, lv);
 
         for i in lim..self.trail.len() {
             let l = self.trail[i];
@@ -320,6 +306,17 @@ impl PropagateIF for AssignStack {
         C: ClauseDBIF,
     {
         let dl = self.decision_level();
+
+        #[cfg(feature = "trail_saving")]
+        if let cc @ Some(_) = self.from_saved_trail(cdb) {
+            self.num_propagation += 1;
+            self.num_conflict += 1;
+            self.num_reconflict += 1;
+            self.dpc_ema.update(self.num_decision);
+            self.ppc_ema.update(self.num_propagation);
+            return cc;
+        }
+
         while let Some(p) = self.trail.get(self.q_head) {
             self.num_propagation += 1;
             self.q_head += 1;
@@ -869,58 +866,6 @@ impl AssignStack {
         {
             self.phase_age = 0;
         }
-    }
-}
-
-#[cfg(feature = "trail_saving")]
-impl AssignStack {
-    fn from_saved_trail<C>(&mut self, cdb: &C) -> Option<ConflictContext>
-    where
-        C: ClauseDBIF,
-    {
-        let q = (0.75 * cdb.derefer(cdb::property::Tf64::DpAverageLBD)).max(4.0) as u16;
-        let dl = self.decision_level();
-        for i in (0..self.trail_saved.len()).rev() {
-            let lit = self.trail_saved[i];
-            let vi = lit.vi();
-            let old_reason = self.reason_saved[vi];
-            match (old_reason, self.assigned(lit)) {
-                (_, Some(true)) => (),
-                // reason refinement by ignoring this dependecy like decision var
-                (AssignReason::Implication(c, l), None) if l == NULL_LIT && q < cdb[c].rank => {
-                    self.trail_saved.truncate(i + 1);
-                    return None;
-                }
-                (AssignReason::Implication(cid, l), None) => {
-                    self.num_repropagation += 1;
-                    self.assign_by_implication(lit, dl, cid, l);
-                }
-                (AssignReason::Implication(cid, link), Some(false)) => {
-                    self.trail_saved.clear();
-                    if link == NULL_LIT {
-                        return Some(ConflictContext { cid, link });
-                    }
-                    if self.reason[vi] == old_reason {
-                        return Some(ConflictContext { cid, link });
-                    } else {
-                        return Some(ConflictContext {
-                            cid,
-                            link: cdb[cid].lit0(),
-                        });
-                    }
-                }
-                (AssignReason::Decision(_), _) => {
-                    self.trail_saved.truncate(i + 1);
-                    return None;
-                }
-                r => panic!("impossible path {:?}", r),
-            }
-        }
-        self.trail_saved.clear();
-        None
-    }
-    pub fn clear_saved_trail(&mut self) {
-        self.trail_saved.clear();
     }
 }
